Cambios generales.

+++ libreria/Funciones.cpp
@@ -8,9 +8,9 @@ float dist(float x1, float y1, float x2, float y2)
     return (distancia);
 }
 
-float area(float r)
+double area(float r)
 {
-    float pi=3.141516;
-    float area=pi*(r*r);
+    double pi=3.141516;
+    double area=pi*(r*r);
     return (area);
 }
